// src/app/model/app-management-model.ts
export interface AppUserManagementModel {
    userId: number | null;
    userName: string | null;
    userPassword: string | null;
    email: string | null;
    mobileNo: string | null;
    userCode: string | null;
    designationCode: string | null;
    db: string | null;
    location: string | null;
}